+++ src/main/java/gov/nist/csd/pm/pap/mysql/MysqlUserDefinedPML.java
@@ -1,7 +1,8 @@
 package gov.nist.csd.pm.pap.mysql;
 
+import gov.nist.csd.pm.pap.UserDefinedPMLStore;
 import gov.nist.csd.pm.policy.UserDefinedPML;
-import gov.nist.csd.pm.policy.exceptions.PMException;
+import gov.nist.csd.pm.policy.exceptions.*;
 import gov.nist.csd.pm.policy.pml.model.expression.Value;
 import gov.nist.csd.pm.policy.pml.statement.FunctionDefinitionStatement;
 import org.apache.commons.lang3.SerializationUtils;
@@ -13,7 +14,7 @@ import java.sql.SQLException;
 import java.util.HashMap;
 import java.util.Map;
 
-public class MysqlUserDefinedPML implements UserDefinedPML {
+public class MysqlUserDefinedPML implements UserDefinedPMLStore {
 
     private MysqlConnection connection;
 
@@ -22,7 +23,10 @@ public class MysqlUserDefinedPML implements UserDefinedPML {
     }
 
     @Override
-    public void createFunction(FunctionDefinitionStatement functionDefinitionStatement) throws MysqlPolicyException {
+    public void createFunction(FunctionDefinitionStatement functionDefinitionStatement)
+    throws PMBackendException, PMLFunctionAlreadyDefinedException {
+        checkCreateFunctionInput(functionDefinitionStatement.getFunctionName());
+
         String sql = """
                 insert into pml_function (name, bytes) values (?,?)
                 """;
@@ -36,7 +40,11 @@ public class MysqlUserDefinedPML implements UserDefinedPML {
     }
 
     @Override
-    public void deleteFunction(String functionName) throws MysqlPolicyException {
+    public void deleteFunction(String functionName) throws PMBackendException {
+        if (!checkDeleteFunctionInput(functionName)) {
+            return;
+        }
+
         String sql = """
                 delete from pml_function where name=?
                 """;
@@ -49,7 +57,7 @@ public class MysqlUserDefinedPML implements UserDefinedPML {
     }
 
     @Override
-    public Map<String, FunctionDefinitionStatement> getFunctions() throws PMException {
+    public Map<String, FunctionDefinitionStatement> getFunctions() throws MysqlPolicyException {
         String sql = """
                 select bytes from pml_function
                 """;
@@ -69,12 +77,14 @@ public class MysqlUserDefinedPML implements UserDefinedPML {
     }
 
     @Override
-    public FunctionDefinitionStatement getFunction(String name) throws PMException {
+    public FunctionDefinitionStatement getFunction(String name)
+    throws PMLFunctionNotDefinedException, PMBackendException {
+        checkGetFunctionInput(name);
+
         String sql = """
                 select bytes from pml_function where name = ?
                 """;
 
-        Map<String, FunctionDefinitionStatement> functionDefinitionStatements = new HashMap<>();
         try(PreparedStatement ps = connection.getConnection().prepareStatement(sql)) {
             ps.setString(1, name);
 
@@ -82,7 +92,6 @@ public class MysqlUserDefinedPML implements UserDefinedPML {
             ResultSet resultSet = ps.executeQuery();
             if (resultSet.next()) {
                 funcDef = SerializationUtils.deserialize(resultSet.getBlob(1).getBinaryStream().readAllBytes());
-                functionDefinitionStatements.put(funcDef.getFunctionName(), funcDef);
             }
 
             resultSet.close();
@@ -94,7 +103,10 @@ public class MysqlUserDefinedPML implements UserDefinedPML {
     }
 
     @Override
-    public void createConstant(String constantName, Value constantValue) throws MysqlPolicyException {
+    public void createConstant(String constantName, Value constantValue)
+    throws PMBackendException, PMLConstantAlreadyDefinedException {
+        checkCreateConstantInput(constantName);
+
         String sql = """
                 insert into pml_constant (name, value) values (?,?)
                 """;
@@ -108,7 +120,11 @@ public class MysqlUserDefinedPML implements UserDefinedPML {
     }
 
     @Override
-    public void deleteConstant(String constName) throws MysqlPolicyException {
+    public void deleteConstant(String constName) throws PMBackendException {
+        if (!checkDeleteConstantInput(constName)) {
+            return;
+        }
+
         String sql = """
                 delete from pml_constant where name=?
                 """;
@@ -121,7 +137,7 @@ public class MysqlUserDefinedPML implements UserDefinedPML {
     }
 
     @Override
-    public Map<String, Value> getConstants() throws PMException {
+    public Map<String, Value> getConstants() throws MysqlPolicyException {
         String sql = """
                 select name, value from pml_constant
                 """;
@@ -142,7 +158,9 @@ public class MysqlUserDefinedPML implements UserDefinedPML {
     }
 
     @Override
-    public Value getConstant(String name) throws PMException {
+    public Value getConstant(String name) throws PMBackendException, PMLConstantNotDefinedException {
+        checkGetConstantInput(name);
+
         String sql = """
                 select value from pml_constant where name=?
                 """;
@@ -160,6 +178,78 @@ public class MysqlUserDefinedPML implements UserDefinedPML {
 
             return value;
         } catch (SQLException | IOException e) {
+            throw new MysqlPolicyException(e.getMessage());
+        }
+    }
+
+    @Override
+    public void checkCreateFunctionInput(String name) throws PMLFunctionAlreadyDefinedException, PMBackendException {
+        String sql = "select name from pml_function where name=?";
+
+        try(PreparedStatement ps = connection.getConnection().prepareStatement(sql)) {
+            ps.setString(1, name);
+
+            ResultSet resultSet = ps.executeQuery();
+            if (resultSet.next()) {
+                throw new PMLFunctionAlreadyDefinedException(name);
+            }
+
+            resultSet.close();
+        } catch (SQLException e) {
+            throw new MysqlPolicyException(e.getMessage());
+        }
+    }
+
+    @Override
+    public void checkGetFunctionInput(String name) throws PMLFunctionNotDefinedException, PMBackendException {
+        String sql = "select name from pml_function where name=?";
+
+        try(PreparedStatement ps = connection.getConnection().prepareStatement(sql)) {
+            ps.setString(1, name);
+
+            ResultSet resultSet = ps.executeQuery();
+            if (!resultSet.next()) {
+                throw new PMLFunctionNotDefinedException(name);
+            }
+
+            resultSet.close();
+        } catch (SQLException e) {
+            throw new MysqlPolicyException(e.getMessage());
+        }
+    }
+
+    @Override
+    public void checkCreateConstantInput(String name) throws PMLConstantAlreadyDefinedException, PMBackendException {
+        String sql = "select name from pml_constant where name=?";
+
+        try(PreparedStatement ps = connection.getConnection().prepareStatement(sql)) {
+            ps.setString(1, name);
+
+            ResultSet resultSet = ps.executeQuery();
+            if (resultSet.next()) {
+                throw new PMLConstantAlreadyDefinedException(name);
+            }
+
+            resultSet.close();
+        } catch (SQLException e) {
+            throw new MysqlPolicyException(e.getMessage());
+        }
+    }
+
+    @Override
+    public void checkGetConstantInput(String name) throws PMLConstantNotDefinedException, PMBackendException {
+        String sql = "select name from pml_constant where name=?";
+
+        try(PreparedStatement ps = connection.getConnection().prepareStatement(sql)) {
+            ps.setString(1, name);
+
+            ResultSet resultSet = ps.executeQuery();
+            if (!resultSet.next()) {
+                throw new PMLConstantNotDefinedException(name);
+            }
+
+            resultSet.close();
+        } catch (SQLException e) {
             throw new MysqlPolicyException(e.getMessage());
         }
     }
